% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-score-summary.R
\name{add-score-summary}
\alias{add-score-summary}
\alias{add_score_summary}
\alias{add_game_summary}
\alias{add_player_summary}
\title{Add score summary to competition results}
\usage{
add_score_summary(cr_data, item, score_fun = NULL, ...)

add_game_summary(cr_data, score_fun = NULL, ...)

add_player_summary(cr_data, score_fun = NULL, ...)
}
\arguments{
\item{cr_data}{Competition results.}

\item{item}{Character vector of columns to group by.}

\item{score_fun}{Function to compute score summary (see Details).}

\item{...}{Additional arguments to be passed to or from methods.}
}
\value{
\code{add_score_summary} returns \code{cr_data} with added
  (by left joining) score summaries.

  \code{add_game_summary} and \code{add_player_summary} are wrappers around
  \code{add_score_summary} with \code{item} equals to "game" and "player"
  accordingly.
}
\description{
This page describes function for adding score summary to competition results
as columns.
}
\details{
\code{cr_data} should be in format suitable for
  \code{\link[=results-longcr]{to_longcr}}.

  If "game" is present in \code{item} but not in column names of
  \code{cr_data} (which can occur in case of
  \code{\link[=results-widecr]{widecr}} format) then every row is treated as
  separate game results. In all other cases every entry of \code{item} should
  be present in column names of \code{cr_data}.

  As in \link{score-summary} one can control the names of the summaries by
  adding prefix stored in \code{prefix} as extra argument.
}
\examples{
cr_data <- data.frame(
  game = rep(1:20, each = 2),
  player = rep(1:10, times = 4),
  score = 31:70,
  season = rep(1:2, each = 20)
)

# Computing summaries.
add_game_summary(cr_data = cr_data, score_fun = score_min_max)
add_player_summary(cr_data = cr_data, score_fun = score_min_max)
add_score_summary(
  cr_data = cr_data, item = c("season", "player"),
  score_fun = score_min_max
)
add_score_summary(
  cr_data = cr_data, item = c("season", "player"),
  score_fun = score_min_max, prefix = "seasonPlayer_"
)

# Using without column "game"
cr_data <- to_widecr(data.frame(
  player1 = 1:10, score1 = 10:19,
  player2 = c(2:10, 1), score2 = 11:20
))
add_game_summary(cr_data = cr_data, score_fun = score_min_max)

}
