% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colley.R
\name{colley}
\alias{colley}
\alias{rate_colley}
\alias{rank_colley}
\title{Colley method}
\usage{
rate_colley(cr_data, players = NULL)

rank_colley(cr_data, players = NULL, ties = c("average", "first", "last",
  "random", "max", "min"), round_digits = 7)
}
\arguments{
\item{cr_data}{Competition results in format ready for
\code{\link[=results-longcr]{as_longcr}}.}

\item{players}{Vector of players for which rating is computed.}

\item{ties}{Value for \code{ties} in \code{\link{round_rank}}.}

\item{round_digits}{Value for \code{round_digits} in
\code{\link{round_rank}}.}
}
\value{
\code{rate_colley} returns a named vector of the Colley rating. The
  mean rating should be 0.5.

  \code{rank_colley} returns a named vector of
  \link[=rating-ranking]{ranking} using \code{\link{round_rank}}.
}
\description{
Functions to compute rating and ranking using Colley method.
}
\details{
This rating method was initially designed for games between two
  players. That is why competition results are transformed into pairwise
  games \code{cr_pair} via \link{to_pairgames}. Data from \code{cr_pair} is
  used in rating computation.

  It is assumed that score is numeric and higher values are better for
  the player.

  Computation is done based only on the matchups between players in
  argument \code{players}. If \code{NULL} then all players present in
  \code{cr_data} are used. \bold{Note} that it isn't necessary for all
  entries in \code{players} to be present in \code{cr_pair} but it might be
  a good idea in order to obtain plausible outcome rating.

  The outline of the Colley method is as follows:
  \enumerate{
    \item Compute Colley matrix: diagonal elements are equal to number of
      games played by certain player \emph{plus 2}, off-diagonal are equal to
      minus number of common games played. This matrix will be the matrix of
      system of linear equations (SLE);
    \item Compute right-hand side of SLE as 1 + 0.5*("number of player's
      wins" - "number of player's losses");
    \item Solve the SLE. The solution is the Colley rating.
  }
}
\examples{
rate_colley(ncaa2005)
rank_colley(ncaa2005)

rate_colley(ncaa2005, players = c("UNC", "Duke", "Miami", "UVA", "VT"))
rate_colley(ncaa2005, players = c("UNC", "Miami", "UVA", "VT"))

}
\references{
Wesley N. Colley (2002) \emph{Colley’s Bias Free College Football
  Ranking Method: The Colley Matrix Explained}. Available online at
  \url{http://www.colleyrankings.com}
}
