% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert-rating}
\alias{convert-rating}
\alias{to_rating_tbl}
\alias{to_rating_vec}
\title{Convert between ratings as named vector and tibble}
\usage{
to_rating_tbl(rating_vec)

to_rating_vec(rating_tbl)
}
\arguments{
\item{rating_vec}{Named vector of ratings.}

\item{rating_tbl}{Tibble with columns 'player' and 'rating'.}
}
\value{
\code{to_rating_tbl} returns a \code{\link[dplyr]{tibble}} with
two columns: 'player' and 'rating'.

\code{to_rating_vec} returns named vector of ratings.
}
\description{
Functions to convert between different forms of ratings.
}
\examples{
rating_vec <- c("pl1" = 1, "pl2" = 10, "pl3" = 100)
rating_tbl <- to_rating_tbl(rating_vec)
rating_tbl
to_rating_vec(rating_tbl)

}
