% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game-score.R
\name{game-score}
\alias{game-score}
\alias{get_game_scores}
\alias{score_game_mean_sd}
\title{Compute game scores}
\usage{
get_game_scores(cr_data, score_game = NULL)

score_game_mean_sd(cr_data)
}
\arguments{
\item{cr_data}{Competition results.}

\item{score_game}{Function to compute game scores (see Details)}
}
\value{
If \code{score_game} is \code{NULL} then \code{get_game_scores}
returns \code{\link[=tbl_df]{tibble}} with one column \code{game}
which has all games in \code{cr_data}. If not \code{NULL} then there will be
extra columns for every game score that \code{score_game} produces
(names are preserved).

\code{score_game_mean_sd} returns a named vector: "meanScore" - mean score
of the game; "sdScore" - standard deviation of scores in the game.
}
\description{
This page describes functionality for computing game scores, i.e. some
measurements (of arbitrary nature) of the game itself.
}
\details{
\code{score_game} is a function that takes results of a
game in \code{\link[=results-longcr]{longcr}} form and produces named
vector of game scores.
}
\examples{
cr_data <- data.frame(
  player = rep(1:5, times = 2),
  game = rep(1:5, each = 2),
  score = 31:40
)

get_game_scores(cr_data = cr_data, score_game = score_game_mean_sd)

}
