% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pf_vec}
\alias{get_pf_vec}
\title{Compute Perron-Frobenius vector}
\usage{
get_pf_vec(mat)
}
\arguments{
\item{mat}{Square matrix}
}
\value{
Perron-Frobenius vector which values sum to 1
}
\description{
Function to compute Perron-Frobenius vector.
}
\keyword{internal}
