% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head-to-head.R
\name{head-to-head-helpers}
\alias{head-to-head-helpers}
\alias{players_drop}
\alias{h2h_fill}
\alias{print.h2h}
\alias{get_players}
\alias{get_matchups}
\title{Helpers for computing Head-to-Head matrices}
\usage{
players_drop(h2h_mat, ...)

h2h_fill(h2h_mat, fill = NA_real_, ...)

\method{print}{h2h}(x, ...)

get_players(cr_data, players = NULL, ...)

get_matchups(cr_data)
}
\arguments{
\item{h2h_mat}{Head-to-Head matrix.}

\item{...}{Additional arguments to be passed to or from methods.}

\item{fill}{Value to fill with.}

\item{x}{Value to print.}

\item{cr_data}{Competition results.}

\item{players}{Vector of players for which Head-to-Head is computed.}
}
\value{
\code{players_drop} removes those rows and column of \code{h2h_mat}
which are filled with \code{NA}s (and therefore the respective players
are absent in original competition results).

\code{h2h_fill} replaces \code{NA}s in \code{h2h_mat} with \code{fill}.

\code{get_players} in case of argument \code{players} is \code{NULL} returns
vector of players present in \code{cr_data}. If not \code{NULL} then returns
its argument \code{players}.

\code{get_matchups} returns a \code{\link[=tbl_df]{tibble}} of all matchups
present in \code{cr_data}. It has following columns:
\describe{
  \item{game}{Game identifier of matchup}
  \item{player1}{Identifier of first player in matchup}
  \item{score1}{Score of the first player in matchup}
  \item{player2}{Identifier of second player in matchup}
  \item{score2}{Score of the second player in matchup}
}
\bold{Note} that matchups are not symmetrical: matchup "player1"-"player2"
is considered different from "player2"-"player1" in order to except more
advanced, not simmetrical Head-to-Head values.
}
\description{
Helpers for computing Head-to-Head matrices
}
\examples{
cr_data <- data.frame(
  game = rep(1:5, each = 3),
  player = rep(1:5, times = 3),
  score = 31:45
)
get_players(cr_data, players = NULL)
get_matchups(cr_data)

}
\seealso{
\link[=head-to-head]{Head-to-Head computing}.
}
