% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head-to-head.R
\name{head-to-head}
\alias{head-to-head}
\alias{h2h_abstract}
\title{Compute Head-to-Head matrix}
\usage{
h2h_abstract(cr_data, h2h_fun, aggr_fun = mean, score_game = NULL,
  players = NULL, absent_players = players_drop, absent_h2h = h2h_fill,
  ...)
}
\arguments{
\item{cr_data}{Competition results.}

\item{h2h_fun}{Head-to-Head function (see Details).}

\item{aggr_fun}{Aggregation function of Head-to-Head values (see Details).}

\item{score_game}{Function for scoring game (see
\link[=game-score]{Computing game scores})}

\item{players}{Vector of players for which Head-to-Head is computed.}

\item{absent_players}{Function which performs actions on Head-to-Head matrix
dealing with players which absent in \code{cr_data}.}

\item{absent_h2h}{Function which performs actions on Head-to-Head matrix
dealing with absent Head-to-Head records for some pairs of players.}

\item{...}{Additional arguments to be passed to or from methods.}
}
\value{
An object of class "h2h" which is a square matrix of Head-to-Head
  values.
}
\description{
This page describes methods of computing Head-to-Head matrices.
}
\details{
\code{h2h_abstract} performs computation of Head-to-Head matrix:
square matrix with number of rows (and columns) equal to number of players
for which it is computed. Head-to-Head values are computed with these steps:
\enumerate{
  \item Compute for every game its scores via \code{score_game}. For details
    see \link[=game-score]{Computing game scores};
  \item Compute for every matchup (pair of players) inside a particular game
    its Head-to-Head value via \code{h2h_fun} based on these players scores
    and game scores. It should accept arguments \code{score1} (for score of
    player1), \code{score2} (for score of player2) and all game scores
    produced by \code{score_game}. \bold{Note} that, for example, matchup
    "player1"-"player2" is considered different from "player2"-"player1" in
    order to except more advanced, not simmetrical Head-to-Head values;
  \item Aggregate for every possible matchup of \code{players} all its
    present Head-to-Head values via \code{aggr_fun}. If some matchup didn't
    occure in \code{cr_data} then \code{NA} is produced;
  \item Perform actions via \code{absent_players} those players' data
    which are absent in \code{cr_data}. For no actions use
    \code{\link{skip_action}};
  \item Perform actions via \code{absent_h2h} on those entries of
    Head-to-Head matrix which matchups didn't occure in \code{cr_data}.
    For no actions use \code{\link{skip_action}}.
}

If \code{score_game} is \code{NULL} then no game scoring is performed.

If \code{players} is \code{NULL} then Head-to-Head matrix is computed for
all present in \code{cr_data} players. \bold{Note} that argument
\code{players} can contain values that are not present in \code{cr_data}:
in this case row and column in Head-to-Head matrix for these values will
contain only \code{NA}s.
}
\examples{
cr_data <- data.frame(
  game = rep(1:5, each = 3),
  player = rep(1:5, times = 3),
  score = 31:45
)
# Compute Head-to-Head matrix with mean score difference.
score_diff <- function(score1, score2) {
  score2 - score1
}
h2h_abstract(
  cr_data = cr_data, h2h_fun = score_diff, aggr_fun = mean,
  score_game = NULL, players = NULL, absent_players = players_drop,
  absent_h2h = h2h_fill
)

}
\seealso{
\link[=head-to-head-helpers]{Head-to-Head helpers}.
}
