% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item-summary-add.R
\name{item-summary-add}
\alias{item-summary-add}
\alias{add_item_summary}
\alias{add_game_summary}
\alias{add_player_summary}
\title{Add item summary to competition results}
\usage{
add_item_summary(cr_data, item, summary_fun = NULL, ...)

add_game_summary(cr_data, summary_fun = NULL, ...)

add_player_summary(cr_data, summary_fun = NULL, ...)
}
\arguments{
\item{cr_data}{Competition results in format ready for
\code{\link[=results-longcr]{to_longcr}}.}

\item{item}{Character vector of columns to group by.}

\item{summary_fun}{Function to compute item summary (see Details).}

\item{...}{Additional arguments to be passed to or from methods.}
}
\value{
\code{add_item_summary} returns \code{cr_data} with added
  (by left joining) item summaries.

  \code{add_game_summary} and \code{add_player_summary} are wrappers around
  \code{add_item_summary} with \code{item} equals to "game" and "player"
  accordingly.
}
\description{
This page describes functions for adding item summary as columns to
competition results.
}
\details{
If "game" is present in \code{item} but not in column names of
  \code{cr_data} (which can occur in case of
  \code{\link[=results-widecr]{widecr}} format) then every row is treated as
  separate game results. In all other cases every entry of \code{item} should
  be present in column names of \code{cr_data}.

  As in \link{item-summary} one can control the names of the summaries by
  adding prefix stored in \code{prefix} as extra argument.
}
\examples{
cr_data <- data.frame(
  game = rep(1:20, each = 2),
  player = rep(1:10, times = 4),
  score = 31:70,
  season = rep(1:2, each = 20)
)

# Computing summaries.
add_game_summary(cr_data = cr_data, summary_fun = summary_min_max_score)
add_player_summary(cr_data = cr_data, summary_fun = summary_min_max_score)
add_item_summary(
  cr_data = cr_data, item = c("season", "player"),
  summary_fun = summary_min_max_score
)
add_item_summary(
  cr_data = cr_data, item = c("season", "player"),
  summary_fun = summary_min_max_score, prefix = "seasonPlayer_"
)

# Using without column "game"
cr_data <- to_widecr(data.frame(
  player1 = 1:10, score1 = 10:19,
  player2 = c(2:10, 1), score2 = 11:20
))
add_game_summary(cr_data = cr_data, summary_fun = summary_min_max_score)

}
\seealso{
\link{item-summary} for description of computing item summary.

\link{item-summary-functions} for implemented \code{summary_fun}s.
}
