% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item-summary-functions.R
\name{item-summary-functions}
\alias{item-summary-functions}
\alias{summary_mean_sd_score}
\alias{summary_min_max_score}
\alias{summary_sum_score}
\title{Specific functions for computing item summary}
\usage{
summary_mean_sd_score(cr_data, prefix = "", ...)

summary_min_max_score(cr_data, prefix = "", ...)

summary_sum_score(cr_data, prefix = "", ...)
}
\arguments{
\item{cr_data}{Competition results.}

\item{prefix}{Possible prefix to add to item summary's names.}

\item{...}{Additional arguments to be passed to or from methods.}
}
\value{
This is a list of currently implemented \code{summary_fun}s:
\describe{
  \item{summary_mean_sd_score}{Returns a named vector (before adding prefix):
    "meanScore" - mean score; "sdScore" - standard deviation of scores;}
  \item{summary_min_max_score}{Returns a named vector (before adding prefix):
    "minScore" - minimum score; "maxScore" - maximum score;}
  \item{summary_sum_score}{Returns a named vector (before adding prefix):
    "sumScore" - sum of all present scores.}
}
}
\description{
This page describes specific functions for computing item summary.
}
\details{
\code{cr_data} should be convertable to
  \code{\link[=results-longcr]{longcr}} form via
  \code{\link[=results-longcr]{to_longcr}}

  Implementations of \code{summary_fun} should be able to take
  \code{prefix} and \code{...} as parameters.
}
\examples{
cr_data <- data.frame(
  game = rep(1:20, each = 2),
  player = rep(1:10, times = 4),
  score = 31:70,
  season = rep(1:2, each = 20)
)

# Computing item summary
summary_mean_sd_score(cr_data)
summary_min_max_score(cr_data)
summary_min_max_score(cr_data, prefix = "total_")

}
\seealso{
\link{item-summary} for description of computing item summary.

\link{item-summary-add} for adding item summary to competition
 results.
}
