% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item-summary.R
\name{item-summary}
\alias{item-summary}
\alias{get_item_summary}
\alias{get_game_summary}
\alias{get_player_summary}
\title{Compute item summary}
\usage{
get_item_summary(cr_data, item, summary_fun = NULL, ...)

get_game_summary(cr_data, summary_fun = NULL, ...)

get_player_summary(cr_data, summary_fun = NULL, ...)
}
\arguments{
\item{cr_data}{Competition results in format ready for
\code{\link[=results-longcr]{to_longcr}}.}

\item{item}{Character vector of columns to group by.}

\item{summary_fun}{Function to compute item summary (see Details).}

\item{...}{Additional arguments to be passed to or from methods.}
}
\value{
If \code{summary_fun} is \code{NULL} then \code{get_item_summary}
  returns a \code{\link[=tbl_df]{tibble}} with columns named as stored in
  argument \code{item} and which has all unique values of particular item
  in \code{cr_data}. If not \code{NULL} then there will be extra columns for
  every summary value that \code{summary_fun} produces.

  \code{get_game_summary} and \code{get_player_summary} are wrappers around
  \code{get_item_summary} with \code{item} equals to "game" and "player"
  accordingly.
}
\description{
This page describes functionality for computing item summary, i.e. some
summary measurements (of arbitrary nature) of item (one or more columns)
present in competition results.
}
\details{
Argument \code{item} defines on which columns grouping is made for
  computing item summary. Basically \code{get_item_summary} applies
  \code{summary_fun} to groups of \code{cr_data} defined by \code{item}.

  \code{summary_fun} is a function that takes competition results of a
  particular item (game, player, their combination, etc.) and produces named
  vector of item summary. Also it should take \code{prefix} and \code{...} as
  argument for easier use. See \link{item-summary-functions} for more
  details.

  One can control the names of the summaries by adding prefix stored in
  \code{prefix} as extra argument.
}
\examples{
cr_data <- data.frame(
  game = rep(1:20, each = 2),
  player = rep(1:10, times = 4),
  score = 31:70,
  season = rep(1:2, each = 20)
)

# Computing summaries.
get_game_summary(cr_data = cr_data, summary_fun = summary_min_max_score)
get_player_summary(cr_data = cr_data, summary_fun = summary_min_max_score)
get_item_summary(
  cr_data = cr_data, item = c("season", "player"),
  summary_fun = summary_min_max_score
)

# Varying prefix of the summary columns.
get_item_summary(
  cr_data = cr_data, item = c("season", "player"),
  summary_fun = summary_mean_sd_score, prefix = "seasonPlayer_"
)

}
\seealso{
\link{item-summary-functions} for implemented \code{summary_fun}s.

\link{item-summary-add} for adding item summary to competition
  results.
}
