% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massey2-fit.R
\name{massey2}
\alias{massey2}
\alias{massey2.default}
\alias{massey2.data.frame}
\alias{massey2.matrix}
\alias{massey2.formula}
\alias{massey2.recipe}
\title{Fit a \code{massey2}}
\usage{
massey2(x, ...)

\method{massey2}{default}(x, ...)

\method{massey2}{data.frame}(x, y, ...)

\method{massey2}{matrix}(x, y, ...)

\method{massey2}{formula}(formula, data, ...)

\method{massey2}{recipe}(x, data, ...)
}
\arguments{
\item{x}{Depending on the context:
\itemize{
\item A \strong{data frame} of predictors.
\item A \strong{matrix} of predictors.
\item A \strong{recipe} specifying a set of preprocessing steps
created from \code{\link[recipes:recipe]{recipes::recipe()}}.
}}

\item{...}{Not currently used, but required for extensibility.}

\item{y}{When \code{x} is a \strong{data frame} or \strong{matrix}, \code{y} is the outcome
specified as:
\itemize{
\item A \strong{data frame} with 1 numeric column.
\item A \strong{matrix} with 1 numeric column.
\item A numeric \strong{vector}.
}}

\item{formula}{A formula specifying the outcome terms on the left-hand side,
and the predictor terms on the right-hand side.}

\item{data}{When a \strong{recipe} or \strong{formula} is used, \code{data} is specified as:
\itemize{
\item A \strong{data frame} containing both the predictors and the outcome.
}}
}
\value{
A \code{massey2} object.
}
\description{
\code{massey2()} fits a model using the Massey Method
}
\examples{
predictors <- mtcars[, -1]
outcome <- mtcars[, 1]

# XY interface
mod <- massey2(predictors, outcome)

# Formula interface
mod2 <- massey2(mpg ~ ., mtcars)

# Recipes interface
library(recipes)
rec <- recipe(mpg ~ ., mtcars)
rec <- step_log(rec, disp)
mod3 <- massey2(rec, mtcars)

}
