% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncaa2005}
\alias{ncaa2005}
\title{Example competition results from 2005 NCAA football season}
\format{An object of class \code{\link[=results-longcr]{longcr}} containing
  information about 10 games.}
\usage{
ncaa2005
}
\description{
\code{ncaa2005} is an example competition results of an isolated group of
Atlantic Coast Conference teams provided in Langville and Meyer book
"Who's #1".
}
\keyword{datasets}
