% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate-colley.R
\name{rate_colley}
\alias{rate_colley}
\title{Colley rating}
\usage{
rate_colley(cr_data, players = NULL)
}
\arguments{
\item{cr_data}{Competition results in format ready for
\code{\link[=results-longcr]{to_longcr}}.}

\item{players}{Vector of players for which rating is computed.}
}
\value{
Named vector of the Colley rating. The sum of all ratings should be
  equal to 0.5*"number of players in the rating".
}
\description{
Function to compute rating using Colley method.
}
\details{
This rating method was initially designed for games between two
  players. That is why competition results are transformed into pairwise
  games \code{cr_pair} with \link{to_pairgames}. Data from \code{cr_pair} is
  used in rating computation.

  Computation is done based only on the matchups between players in
  argument \code{players}. If \code{NULL} then all players present in
  \code{cr_data} are used. \bold{Note} that it isn't necassary for all
  entries in \code{players} to be present in \code{cr_pair} but it might be
  a good idea in order to obtain plausible outcome rating.

  The outline of the Colley method is as follows:
  \enumerate{
    \item Compute Colley matrix: diagonal elements are equal to number of
      games played by certain player \emph{plus 2}, off-diagonal are equal to
      minus number of common games played. This matrix will be the matrix of
      system of linear equations (SLE);
    \item Compute right-hand side of SLE as 1 + 0.5*("number of player's
      wins" - "number of player's losses");
    \item Solve the SLE. The solution is the Colley rating.
  }
}
\examples{
rate_colley(ncaa2005)
rate_colley(ncaa2005, players = c("UNC", "Duke", "Miami", "UVA", "VT"))
rate_colley(ncaa2005, players = c("UNC", "Miami", "UVA", "VT"))

}
