% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate-massey.R
\name{rate_massey}
\alias{rate_massey}
\title{Massey rating}
\usage{
rate_massey(cr_data, players = NULL)
}
\arguments{
\item{cr_data}{Competition results in format ready for
\code{\link[=results-longcr]{to_longcr}}.}

\item{players}{Vector of players for which rating is computed.}
}
\value{
Named vector of the Massey rating. The sum of all ratings should be
  equal to 0.
}
\description{
Function to compute rating using Massey method.
}
\details{
This rating method was initially designed for games between two
  players. That is why competition results are transformed into pairwise
  games \code{cr_pair} via \link{to_pairgames}. Data from \code{cr_pair} is
  used in rating computation.

  It is assumed that score is numeric and higher values are better for
  the player.

  Computation is done based only on the games between players in argument
  \code{players}. If \code{NULL} then all players present in \code{cr_pair}
  are used. \bold{Note} that all \code{players} should be present in
  \code{cr_pair} because otherwise there will be an error during solving
  linear system described below. Message is given if there are players absent
  in \code{cr_pair}.

  The outline of Massey rating method is as follows:
  \enumerate{
    \item Compute Massey matrix: diagonal elements are equal to number of
      games played by certain player, off-diagonal are equal to minus number
      of common games played. This matrix will be the matrix of system of
      linear equations (SLE);
    \item Compute right-hand side of SLE as cumulative score differences of
      players, i.e. sum of all scores \emph{for} the player minus sum of all
      scores \emph{against} the player;
    \item Make adjustment for solvability of SLE. Modify the last row of
      Massey matrix so that all its cells are equal to 1. Also change the
      last cell in right-hand side to 0. This adjustment ensures that sum of
      all outcome ratings will be 0;
    \item Solve the SLE. The solution is the Massey rating.
  }
}
\examples{
rate_massey(ncaa2005)
rate_massey(ncaa2005, players = c("UNC", "Duke", "Miami", "UVA", "VT"))
rate_massey(ncaa2005, players = c("UNC", "Miami", "UVA", "VT"))

}
