% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating-ranking.R
\name{rating-ranking}
\alias{rating-ranking}
\title{Definition of Rating and Ranking}
\description{
\itemize{
\item \strong{Rating} - List (in the ordinary sense) of numerical values, one for each
player, or the numerical value itself. Its interpretation depends on rating
method: either bigger value indicates better player performance or otherwise.
\item \strong{Ranking} - Rank-ordered list (in the ordinary sense) of players: rank 1
indicates player with best performance.
}
}
