% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-summary-functions.R
\name{score-summary-functions}
\alias{score-summary-functions}
\alias{score_mean_sd}
\alias{score_min_max}
\title{Specific functions for computing score summary}
\usage{
score_mean_sd(cr_data, prefix = "", ...)

score_min_max(cr_data, prefix = "", ...)
}
\arguments{
\item{cr_data}{Competition results.}

\item{prefix}{Possible prefix to add to score summary's names.}

\item{...}{Additional arguments to be passed to or from methods.}
}
\value{
This is a list of currently implemented \code{score_fun}s:
\describe{
  \item{score_mean_sd}{Returns a named vector (before adding prefix):
  "meanScore" - mean score; "sdScore" - standard deviation of scores;}
  \item{score_min_max}{Returns a named vector (before adding prefix):
  "minScore" - minimum score; "maxScore" - maximum score.}
}
}
\description{
This page describes specific functions for computing score summary.
}
\details{
\code{cr_data} should be convertable to
  \code{\link[=results-longcr]{longcr}} form via
  \code{\link[=results-longcr]{to_longcr}}

  Implementations of \code{score_fun} should be able to take
  \code{prefix} and \code{...} as parameters.
}
\examples{
cr_data <- data.frame(
  game = rep(1:20, each = 2),
  player = rep(1:10, times = 4),
  score = 31:70,
  season = rep(1:2, each = 20)
)

# Computing score summary
score_mean_sd(cr_data)
score_min_max(cr_data)
score_min_max(cr_data, prefix = "total_")

}
