% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{skip_action}
\alias{skip_action}
\title{Skip some action}
\usage{
skip_action(x, ...)
}
\arguments{
\item{x}{Input object}

\item{...}{Additional arguments to be passed to or from methods.}
}
\value{
An exact copy of \code{x}.
}
\description{
Skip some action
}
\examples{
identical(skip_action(1), 1)

}
