% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head-to-head-functions.R
\name{head-to-head-functions}
\alias{head-to-head-functions}
\alias{h2h_mean_score_diff}
\alias{h2h_mean_score}
\alias{h2h_sum_score_diff}
\alias{h2h_sum_score}
\alias{h2h_num_wins}
\title{Specific functions for computing Head-to-Head matrices}
\usage{
h2h_mean_score_diff(matchup_data, ...)

h2h_mean_score(matchup_data, ...)

h2h_sum_score_diff(matchup_data, ...)

h2h_sum_score(matchup_data, ...)

h2h_num_wins(matchup_data, half_for_draw = FALSE, ...)
}
\arguments{
\item{matchup_data}{Data of matchups described in \link{head-to-head} as
input for specific \code{h2h_fun}.}

\item{...}{Additional arguments to be passed to or from methods.}

\item{half_for_draw}{Whether to count half points for draws in
\code{h2h_num_wins}.}
}
\value{
All \code{h2h_*} are implementations of \code{h2h_fun} and return a
  single Head-to-Head value.
}
\description{
Specific functions for computing Head-to-Head matrices.
}
\details{
This is a list of currently implemented \code{h2h_fun}s:
\describe{
  \item{h2h_mean_score_diff}{Computes mean score difference of \code{player2}
    compared to \code{player1};}
  \item{h2h_mean_score}{Computes mean score of \code{player2};}
  \item{h2h_sum_score_diff}{Computes sum of score differences of
    \code{player2} compared to \code{player1};}
  \item{h2h_sum_score}{Computes sum of scores of \code{player2};}
  \item{h2h_num_wins}{Computes number of matchups \code{player2} scored
    \bold{more} than \code{player1}. If \code{half_for_draw} is \code{TRUE}
    then it also adds half number of matchups where they had draw.
    \bold{Note} that it excludes matchups of the players with themselves.}
}
}
\examples{
# Initializing data
cr_data <- data.frame(
  game = rep(1:5, each = 2),
  player = rep(1:5, times = 2),
  score = c(31:39, 39)
)
matchup_data <- get_matchups(cr_data)

# Computing Head-to-Head matrices
get_h2h(cr_data, h2h_num_wins)
get_h2h(cr_data, h2h_num_wins, half_for_draw = TRUE)
get_h2h(cr_data, h2h_num_wins, half_for_draw = TRUE, fill = 0)

# Computing Head-to-Head values based on matchups
h2h_mean_score_diff(matchup_data[2, ])
h2h_mean_score_diff(matchup_data[19, ])

}
\seealso{
\link{head-to-head} Abstract computing of Head-to-Head matrices.

\link{head-to-head-helpers} Head-to-Head helpers.
}
