% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item-summary-functions.R
\name{item-summary-functions}
\alias{item-summary-functions}
\alias{summary_mean_sd_score}
\alias{summary_min_max_score}
\alias{summary_sum_score}
\alias{summary_num_games}
\alias{summary_num_players}
\title{Specific functions for computing item summary}
\usage{
summary_mean_sd_score(cr_data, prefix = "", ...)

summary_min_max_score(cr_data, prefix = "", ...)

summary_sum_score(cr_data, prefix = "", ...)

summary_num_games(cr_data, prefix = "", ...)

summary_num_players(cr_data, prefix = "", ...)
}
\arguments{
\item{cr_data}{Competition results in format ready for
\code{\link[=results-longcr]{to_longcr}}.}

\item{prefix}{Possible prefix to add to item summary's names.}

\item{...}{Additional arguments to be passed to or from methods.}
}
\value{
All \code{summary_fun}s return a named vector of item summaries
  (names are given before adding prefix). This is a list of currently
  implemented \code{summary_fun}s:
\describe{
  \item{summary_mean_sd_score}{Returns a vector:
    \code{meanScore} - mean score;
    \code{sdScore} - standard deviation of scores;}
  \item{summary_min_max_score}{Returns a vector:
    \code{minScore} - minimum score; \code{maxScore} - maximum score;}
  \item{summary_sum_score}{Returns a vector:
    \code{sumScore} - sum of all present scores;}
  \item{summary_num_games}{Returns a vector:
    \code{numGames} - number of unique games;}
  \item{summary_num_players}{Returns a vector:
    \code{numPlayers} - number of unique players.}
}
}
\description{
This page describes specific functions for computing item summary.
}
\details{
Implementations of \code{summary_fun} should be able to take
  \code{prefix} and \code{...} as parameters.
}
\examples{
cr_data <- data.frame(
  game = rep(1:20, each = 2),
  player = rep(1:10, times = 4),
  score = 31:70,
  season = rep(1:2, each = 20)
)

# Computing item summary
summary_mean_sd_score(cr_data)
summary_min_max_score(cr_data)
summary_min_max_score(cr_data, prefix = "total_")

}
\seealso{
\link{item-summary} for description of computing item summary.

\link{item-summary-add} for adding item summary to competition
 results.
}
