% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate-keener.R
\name{keener}
\alias{keener}
\alias{rate_keener}
\alias{skew_keener}
\alias{normalize_keener}
\title{Keener rating}
\usage{
rate_keener(cr_data, h2h_fun, players = NULL, force_nonneg_h2h = TRUE,
  skew_fun = skew_keener, normalize_fun = normalize_keener, eps = 0.001,
  ...)

skew_keener(x, ...)

normalize_keener(h2h_mat, cr_data, ...)
}
\arguments{
\item{cr_data}{Competition results in format ready for
\code{\link[=results-longcr]{to_longcr}}.}

\item{h2h_fun}{Head-to-Head function to compute Head-to-Head matrix.}

\item{players}{Vector of players for which rating is computed.}

\item{force_nonneg_h2h}{Whether to force nonnegative values in Head-to-Head
matrix.}

\item{skew_fun}{Skew function.}

\item{normalize_fun}{Normalization function.}

\item{eps}{Coefficient for forcing irreducibility.}

\item{...}{Additional arguments to be passed to methods.}

\item{x}{Argument for \code{skew_keener}.}

\item{h2h_mat}{Argument for \code{normalize_keener}.}
}
\value{
Named vector of the Keener rating. The sum of all ratings should be
  equal to 1.
}
\description{
Function to compute rating using Keener method.
}
\details{
Keener rating method is based on Head-to-Head matrix of the
  competition results. Therefore it can be used for competitions with
  variable number of players. Its algorithm is as follows:
\enumerate{
  \item Compute Head-to-Head matrix of competition results for \code{players}
    with \code{h2h_fun} via \code{\link{get_h2h}}. Ensure that there are no
    \code{NA}s. Absent in \code{cr_data} players are dropped. If
    \code{force_nonneg_h2h} is \code{TRUE} then the minimum value is
    subtracted (in case some Head-to-Head value is strictly negative).
    \bold{Note} that Keener method is designed for Head-to-Head matrix with
    the following property: the more value in row \bold{i} and column
    \bold{j} the better player from row \bold{i} performed than player from
    column \bold{j}. Extra argument \code{transpose} for
    \code{\link{get_h2h}} can be used to ensure that;
  \item Update raw Head-to-Head values (denoted as S) with the
    pair-normalization: a_{ij} = (S_ij + 1) / (S_ij + S_ji + 2). This step
    should make comparing different players more reasonable;
  \item Skew Head-to-Head values with applying \code{skew_fun} to them.
    \code{skew_fun} should take numeric vector as first argument and
    \code{...} as second. It should return skewed vector. The default skew
    function is \code{skew_keener}. This step should make abnormal results
    not very abnormal. To omit this step use \code{\link{skip_action}};
  \item Normalize Head-to-Head values with \code{normalize_fun} using
    \code{cr_data}. \code{normalize_fun} should take Head-to-Head matrix as
    the first argument, \code{cr_data} as second and \code{...} as third. It
    should return normalized matrix. The default normalization is
    \code{normalize_keener} which divides Head-to-Head value of
    'player1'-'player2' matchup by the number of games played by 'player1'.
    This step should take into account possibly not equal number of games
    played by players. To omit this step use \code{\link{skip_action}};
  \item Add small value to Head-to-Head matrix to ensure its irreducibility.
    If all values are strictly positive then this step is omitted. In other
    case small value is computed as the smallest non-zero Head-to-Head value
    multiplied by \code{eps}. This step is done to ensure applicability of
    Perron-Frobenius theorem;
  \item Compute Perron-Frobenius vector of the resultant matrix, i.e. the
    strictly positive real eigenvector (which values sum to 1) for eigenvalue
    (which is real) of the maximum absolute value. This vector is Keener
    rating vector.
}
}
\examples{
# Use transpose = TRUE is correct in this case
rate_keener(ncaa2005, h2h_sum_score, transpose = TRUE)

# Impact of skewing
rate_keener(ncaa2005, h2h_sum_score, transpose = TRUE,
            skew_fun = skip_action)

# Impact of normalization. Use fill = 0 to ensure no NA after get_h2h
rate_keener(ncaa2005[-(1:2), ], h2h_sum_score,
            transpose = TRUE, fill = 0)
rate_keener(ncaa2005[-(1:2), ], h2h_sum_score,
            normalize_fun = skip_action,
            transpose = TRUE, fill = 0)

}
\references{
James P. Keener (1993) \emph{The Perron-Frobenius theorem and the
  ranking of football teams}. SIAM Review, 35(1):80–93, 1993.
}
