% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-widecr.R
\name{results-widecr}
\alias{results-widecr}
\alias{is_widecr}
\alias{to_widecr}
\title{Wide format of competition results}
\usage{
is_widecr(cr_data)

to_widecr(cr_data, repair = TRUE, ...)
}
\arguments{
\item{cr_data}{Data of competition results (convertable to tabular).}

\item{repair}{Whether to repair input.}

\item{...}{Additional arguments to be passed to or from methods.}
}
\value{
\code{is_widecr} returns TRUE if its argument is appropriate object
  of class \code{widecr}.

\code{to_widecr} returns an object of class \code{widecr}.
}
\description{
Functions for converting data of competition results to wide format.
}
\details{
\code{to_widecr} is S3 method for converting data to \code{widecr}.
When using default method if \code{repair = TRUE} it tries to fix possible
problems with the following actions:
\itemize{
  \item Detect columns with names containing "player" or "score".
    All other columns are treated as "extra";
  \item Extract first occurence of "player" or "score" from names of detected
    columns. Everything after extracted word is treated as identifier of
    "player"-"score" pair;
  \item Convert these identifiers to numeric form with
    \code{as.integer(as.factor(...))};
  \item Convert identifiers once again to character form with possible
    leading zeros (to account for R standard string ordering);
  \item Spread pairs to appropriate columns adding columns with
    \code{NA_integer_} if they were not present in original data.
}
Note that the order (and numeration) of pairs can change.

If \code{repair} is \code{FALSE} it converts \code{cr_data} to
\code{\link[=tbl_df]{tibble}} and adds \code{widecr} class to it.

When applying \code{to_widecr} to \code{longcr} object, convertion is made:
\itemize{
  \item All columns except "game", "player" and "score" are dropped;
  \item Convertion from long to wide format is made. The number of
    "player"-"score" pairs is taken as the maximum number of players in game.
    If not all games are played between the same number of players then there
    will be \code{NA}'s in some pairs.
    Column \code{game} is preserved in output.
}

For appropriate \code{widecr} objects \code{to_widecr} returns its input.
}
\section{Wide format of competition results}{

It is assumed that competition consists from multiple games (matches,
comparisons, etc.). One game can consist only from \bold{constant} number
of players. Inside a game all players are treated equally.
In every game every player has some score: the value of arbitrary nature
that сharacterizes player's performance (in most cases it is some numeric
value).

\code{widecr} inherits from \code{\link[=tbl_df]{tibble}}. Data should be
orginized in pairs of columns "player"-"score". For example: player1, score1,
player2, score2. Order doesn't matter. Extra columns are allowed.

To account for R standard string ordering, identifiers of pairs should be
formatted with possible leading zeros. For example: player01, score01, ...,
player10, score10.

Column \code{game} for game identifier is optional. If present it will
be used in convertion to \code{longcr} format via \code{\link{to_longcr}}.
}

\examples{
cr_data <- data.frame(
  playerA = 1:10,
  playerB = 2:11,
  scoreC = 11:20,
  scoreB = 12:21,
  scoreA = 13:22,
  otherColumn =  101:110
)
cr_data_wide <- to_widecr(cr_data, repair = TRUE)
is_widecr(cr_data_wide)

}
\seealso{
\link[=results-longcr]{longcr} for long format.
}
